remove libc_stat.h
