+++ components/dfs/filesystems/romfs/dfs_romfs.c
@@ -319,4 +319,3 @@ int dfs_romfs_init(void)
     return 0;
 }
 INIT_COMPONENT_EXPORT(dfs_romfs_init);
-
