+++ components/libc/compilers/newlib/stdio.c
@@ -9,7 +9,7 @@
  */
 #include <stdio.h>
 #include <stdlib.h>
-
+#include <fcntl.h>
 #include <rtthread.h>
 #include "libc.h"
