+++ components/net/lwip-2.0.3/src/lwipopts.h
@@ -572,6 +572,9 @@
 #ifndef LWIP_SOCKET
 #define LWIP_SOCKET                     1
 #endif
+#ifdef LWIP_SOCKET
+#include <fcntl.h>
+#endif
 
 /*
  * LWIP_COMPAT_SOCKETS==1: Enable BSD-style sockets functions names.
