+++ components/net/lwip-2.1.2/src/lwipopts.h
@@ -579,6 +579,9 @@
 #ifndef LWIP_SOCKET
 #define LWIP_SOCKET                     1
 #endif
+#ifdef LWIP_SOCKET
+#include <fcntl.h>
+#endif
 
 /*
  * LWIP_COMPAT_SOCKETS==1: Enable BSD-style sockets functions names.
