+++ components/net/netdev/src/netdev_ipaddr.c
@@ -210,13 +210,13 @@ char *netdev_ip4addr_ntoa_r(const ip4_addr_t *addr, char *buf, int buflen)
         {
             if (len++ >= buflen)
             {
-                return NULL;
+                return RT_NULL;
             }
             *rp++ = inv[i];
         }
         if (len++ >= buflen)
         {
-            return NULL;
+            return RT_NULL;
         }
         *rp++ = '.';
         ap++;
@@ -556,11 +556,11 @@ netdev_inet_ntop(int af, const void *src, char *dst, int32_t size)
 #define AF_INET         2
 #define AF_INET6        10
 
-    const char *ret = NULL;
+    const char *ret = RT_NULL;
     int size_int = (int)size;
     if (size_int < 0)
     {
-        return NULL;
+        return RT_NULL;
     }
     switch (af)
     {
