+++ components/net/sal_socket/socket/net_netdb.c
@@ -10,7 +10,7 @@
  */
 
 #include <rtthread.h>
-
+#include <stddef.h>
 #include <netdb.h>
 
 struct hostent *gethostbyname(const char *name)
